43: created Today view component
